requirements.txt
Django==2.2.10
djangorestframework==3.12.3
psycopg2-binary==2.8.6
